CHPS0705/requirements.txt
# Reconnaissance vocale
openai-whisper
sounddevice
soundfile
numpy
scipy

# Analyse IA avec Ollama
requests

# Utilitaires
colorama
